change view model, change constuctor detailedheroscreen
